requirements.txt
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
googletrans==4.0.0rc1
greenlet==3.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
